requirements.txt
numpy
pandas
scikit-learn
faker
pillow
reportlab
openpyxl
fastapi
streamlit
langchain
langgraph
chromadb
pyyaml
ipykernel